# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: USE_NEXT API
  version: 0.0.1
servers:
  - url: /api
...
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Returns the user of the current session
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
...
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    UserInfo:
      type: object
      required: [username]
      properties:
        username:
          type: string
...
    Problem:
      type: object
      required: [status, title]
      properties:
        status:
          type: integer
        title:
          type: string
        detail:
          type: string
          maxLength: 500
          description: Plain-text, sanitised explanation; never contains stack traces, paths or class names